traing activity and ai remodel
