+++ api/ai/repo_scoring_service.py
@@ -1,9 +1,10 @@
 import logging
+import datetime
 from typing import Dict, List, Any
 from ai.type_analyzer import FileTypeAnalyzer
-from ai.semantic_scorer import SemanticScorer
 from config.fine_tuning import SemanticModel
-import datetime
+import numpy as np
+from data_filter import extract_language_terms, technical_terms_structured
 
 logger = logging.getLogger('portfolio.api')
 
@@ -16,7 +17,6 @@ class RepoScoringService:
     def __init__(self, username: str = None):
         """Initialize the repository scoring service with required components."""
         self.semantic_model = SemanticModel()
-        self.semantic_scorer = SemanticScorer(self.semantic_model)
         self.file_type_analyzer = FileTypeAnalyzer()
         self.username = username
         
@@ -113,3 +113,138 @@ class RepoScoringService:
             "total_relevance_score": total_score,
             "scoring_timestamp": datetime.datetime.now().isoformat()
         }
+        
+    def score_context_similarity(self, query: str, repo_bundle: Dict) -> float:
+        """
+        Scores the similarity between the query and the repository context using semantic embeddings.
+        """
+        context_str = self.flatten_repo_context_to_natural_language(repo_bundle)
+        if not context_str or context_str.strip() == "" or "None" in context_str:
+            logger.debug(f"Skipping context scoring for {repo_bundle.get('name', 'Unknown')} due to empty or meaningless context string.")
+            return 0.0
+
+        # Encode with whitening + L2 normalization for better spread
+        q_emb = self.semantic_model.encode([query], apply_whitening=True, normalize=True)
+        c_emb = self.semantic_model.encode([context_str], apply_whitening=True, normalize=True)
+
+        # Dot product equals cosine for normalized vectors
+        similarity = float(np.dot(q_emb[0], c_emb[0]))
+
+        project_name = repo_bundle.get("repoContext", {}).get("project_identity", {}).get("name")
+        logger.debug(f"Context similarity score for: '{project_name}'= {similarity}")
+
+        return similarity
+
+    def score_language_matches(self, query: str, repo_languages: Dict) -> float:
+        """
+        Returns a normalized score [0, 1] based on the proportion of query language matches
+        to the number of query language terms. Ignores size.
+        """
+        query_terms = [t.lower() for t in extract_language_terms(query)]
+        if not query_terms:
+            return 0.0
+        repo_langs = [lang.lower() for lang in repo_languages.keys()]
+        matches = set(query_terms) & set(repo_langs)
+        score = len(matches) / len(query_terms)
+        logger.debug(f"Language match score for query '{query}': {score} (matches: {matches})")
+        return min(score, 1.0)
+
+    def score_language_size(self, query: str, repo_languages: Dict) -> float:
+        """
+        Returns a normalized score [0, 1] based on the total size of matching languages
+        divided by the total size of all languages in the repo.
+        """
+        query_terms = [t.lower() for t in extract_language_terms(query)]
+        if not query_terms or not repo_languages:
+            return 0.0
+        total_size = sum(repo_languages.values())
+        if total_size == 0:
+            return 0.0
+        match_size = sum(size for lang, size in repo_languages.items() if lang.lower() in query_terms)
+        logger.debug(f"Language size score for query '{query}': {match_size}/{total_size} = {match_size / total_size if total_size > 0 else 0.0}")
+        score = match_size / total_size
+        return min(score, 1.0)
+
+    def aggregate_scores(self, context_score: float, language_score: float, type_score: float) -> float:
+        return (context_score * 0.4) + (language_score * 0.3) + (type_score * 0.3)
+    
+    
+    def flatten_repo_context_to_natural_language(self, repo_bundle: Dict) -> str:
+        """
+        Converts the repository context bundle (including .repo-context.json, README.md,
+        SKILLS-INDEX.md, ARCHITECTURE.md) into a natural-language-like paragraph
+        for sentence-transformer embedding and fine-tuning.
+    
+        Note: This method is similar to SemanticModel._flatten_repo_bundle_for_training
+        but maintained separately to allow flexible field inclusion. Keep the core
+        fields in sync with the corresponding method in fine_tuning.py.
+
+        Args:
+            repo_bundle (Dict): Repository context bundle.
+
+        Returns:
+            str: Flattened natural-language representation of the repository context.
+        """
+        lines = []
+
+        # Extract structured fields from repo-context.json
+        identity = repo_bundle.get("repoContext", {}).get("project_identity", {})
+        tech_stack = repo_bundle.get("repoContext", {}).get("tech_stack", {})
+        # skills = repo_bundle.get("repoContext", {}).get("skill_manifest", {})
+        # outcomes = repo_bundle.get("repoContext", {}).get("outcomes", {})
+        # metadata = repo_bundle.get("repoContext", {}).get("metadata", {})
+        # assessment = repo_bundle.get("repoContext", {}).get("assessment", {})
+
+        # Build natural language representation
+        if identity.get('name'):
+            lines.append(f"Project Name: {identity['name']}.")
+        if identity.get('description'):
+            lines.append(f"Description: {identity['description']}.")
+        if identity.get('type'):
+            lines.append(f"Type: {identity['type']}.")
+        if identity.get('scope'):
+            lines.append(f"Scope: {identity['scope']}.")
+
+        if tech_stack.get("primary"):
+            lines.append(f"Primary technologies include {', '.join(tech_stack['primary'])}.")
+        if tech_stack.get("secondary"):
+            lines.append(f"Secondary tools include {', '.join(tech_stack['secondary'])}.")
+        if tech_stack.get("key_libraries"):
+            lines.append(f"Key libraries: {', '.join(tech_stack['key_libraries'])}.")
+        if tech_stack.get("development_tools"):
+            lines.append(f"Development tools used: {', '.join(tech_stack['development_tools'])}.")
+
+        # if skills.get("technical"):
+        #     lines.append(f"Technical skills demonstrated include {', '.join(skills['technical'])}.")
+        # if skills.get("domain"):
+        #     lines.append(f"Domain-specific knowledge areas: {', '.join(skills['domain'])}.")
+        # if skills.get("competency_level"):
+        #     lines.append(f"Competency level: {skills['competency_level']}.")
+
+        # if outcomes.get("deliverables"):
+        #     lines.append(f"Deliverables include {', '.join(outcomes['deliverables'])}.")
+        # if outcomes.get("skills_acquired"):
+        #     lines.append(f"Skills acquired: {', '.join(outcomes['skills_acquired'])}.")
+        # if outcomes.get("primary"):
+        #     lines.append(f"Primary outcomes: {', '.join(outcomes['primary'])}.")
+
+        # if assessment.get("difficulty"):
+        #     lines.append(f"Difficulty level: {assessment['difficulty']}.")
+        # if assessment.get("evaluation_criteria"):
+        #     lines.append(f"Evaluation criteria: {', '.join(assessment['evaluation_criteria'])}.")
+
+        # if metadata.get("tags"):
+        #     lines.append(f"Tags: {', '.join(metadata['tags'])}.")
+        # if metadata.get("maintainer"):
+        #     lines.append(f"Maintainer: {metadata['maintainer']}.")
+        # if metadata.get("license"):
+        #     lines.append(f"License: {metadata['license']}.")
+
+        # # Add README, SKILLS-INDEX, and ARCHITECTURE content
+        # for key in ["readme", "skills_index", "architecture"]:
+        #     content = repo_bundle.get(key)
+        #     if content:
+        #         lines.append(f"{key.capitalize()}: {content.strip()}")
+
+        flattened_context = "\n".join(lines)
+        return flattened_context
